% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{add_noisy_replicates}
\alias{add_noisy_replicates}
\title{Add replicates with some noise to an midata object where experiments are not repeated (and probably noise free also)
Returns a new midata object where the mids per peak per experiment are repeated with a defined level of noise
Updates all related fields in midata such as avg_mids, exp_index, etc.}
\usage{
add_noisy_replicates(midata, stdev, nr_replicate)
}
\arguments{
\item{midata}{An MIData object}

\item{stdev}{The desired standard deviation to generate dirichlet distribution
(see the random_mid() function for more details on the dirichlet distribution)
and return a valid MID vector of the same size.}

\item{nr_replicate}{Number of noisy replicates to generate}
}
\description{
This is useful when the midata object contains simulated MIDs without replicates and noise,
and to add replicates we introduce a level of noise defined by the stdev parameter.
}
