% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{apply_no_m0}
\alias{apply_no_m0}
\title{Apply a function on MIDs x,y after removing M+0}
\usage{
apply_no_m0(f, x, y)
}
\arguments{
\item{f}{function to apply}

\item{x}{a vector}

\item{y}{a vector}
}
\value{
the result of calling f
}
\description{
This simply removes the first component of each vector before applying f
}
