% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{assign_list}
\alias{assign_list}
\title{Unique object used to assign to list elements}
\format{
An object of class \code{AssignList} of length 1.
}
\usage{
assign_list
}
\description{
Unique object used to assign to list elements
}
\keyword{datasets}
