% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{best_nonempty}
\alias{best_nonempty}
\title{Apply a selection function to x while handling the "index" attribute}
\usage{
best_nonempty(x, select)
}
\arguments{
\item{x}{A vector of values. If x has an "index"
attribute, it must be a vector of same length as x.}

\item{select}{A function selecting a value from x, returning its index}
}
\value{
The value in x chosen by select, or NA if x is empty. If x has an "index"
attribute, the element corresponding to the selected value is added as
an "index" attribute to the result.
}
\description{
NOTE: I'm not convinced we should return NA in
case of empty lists; see ?max for reasons for using +/-Inf
Keeping this for compatibility for now
}
