% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{bray_curtis_distance}
\alias{bray_curtis_distance}
\title{The Bray-Curtis distance}
\usage{
bray_curtis_distance(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the Bray-Curtis distance between x and y
}
\description{
The Bray-Curtis distance
}
