% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{c13correct_cols}
\alias{c13correct_cols}
\title{Correct an MID matrix for naturally occurring isotopes}
\usage{
c13correct_cols(mids, p = natural_13C_fraction, constraint = TRUE)
}
\arguments{
\item{mids}{A matrix whose columns are MID vectors}

\item{p}{The heavy atom natural abundance}

\item{constraint}{whether to constrain sum of corrected vectors to 1}
}
\description{
Correct an MID matrix for naturally occurring isotopes
}
