% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calc.cosine}
\alias{calc.cosine}
\title{Cosine similarity between two vectors of the same length.}
\usage{
calc.cosine(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the cosine similarity, or NA if either x or y is a zero vector
}
\description{
Cosine similarity between two vectors of the same length.
}
