% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold-standard.R
\name{calc_accuracy}
\alias{calc_accuracy}
\title{Compute accuracy measures element-wise for two binary matrices}
\usage{
calc_accuracy(true_matrix, predicted_matrix)
}
\arguments{
\item{true_matrix}{Binary matrix where 1 indicates true}

\item{predicted_matrix}{Binary matrix where 1 indicates a positive prediction}
}
\value{
A list with elements $tpr, $fpr, $precision, $fdr, $f1_score
}
\description{
Compute accuracy measures element-wise for two binary matrices
}
