% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{canberra_distance}
\alias{canberra_distance}
\title{The Canberra distance}
\usage{
canberra_distance(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the Canberra distance between x and y
}
\description{
The Canberra distance
}
