% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{censor_false_mi}
\alias{censor_false_mi}
\title{Censor false mass isotopomers}
\usage{
censor_false_mi(mi_data, threshold = 0.03, min_experiments = 1)
}
\arguments{
\item{mi_data}{An MIData object}

\item{threshold}{MI fraction threshold (after 13C correction)}

\item{min_experiment}{Minimum number of experiments where an MI is observed
to be considered fasle}
}
\value{
A new MIData object with false MIs censored
}
\description{
Find mass isotopomers whose fraction is above the given threshold
after correction for natural 13C in at least the given number of
experiments (inclusive), set them to zero, and renormalize
}
