% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{combine_sqrt_sum}
\alias{combine_sqrt_sum}
\title{Combines pairwise matrices across multiple experiments by their squared sum.
Note that this cannot be used if the middle metabolite matrix and experiment index matrix are to be tracked.}
\usage{
combine_sqrt_sum(pairwise_matrices)
}
\arguments{
\item{pairwise_matrices}{a list of pairwise matrices across different experiments}
}
\description{
Combines pairwise matrices across multiple experiments by their squared sum.
Note that this cannot be used if the middle metabolite matrix and experiment index matrix are to be tracked.
}
