% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{compare_formula_diff_to_list}
\alias{compare_formula_diff_to_list}
\title{Compares a formula to a list of formulas, and returns the match if there is any}
\usage{
compare_formula_diff_to_list(formula_diff, allowed_reactions)
}
\description{
Compares a formula to a list of formulas, and returns the match if there is any
}
