% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{compare_mass_diff_to_list}
\alias{compare_mass_diff_to_list}
\title{Compares a mass to a list of masses, and returns the match if there is any within the given error range}
\usage{
compare_mass_diff_to_list(mass_diff, allowed_masses, tol = 10)
}
\description{
Compares a mass to a list of masses, and returns the match if there is any within the given error range
}
