% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold-standard.R
\name{continuous_accuracy}
\alias{continuous_accuracy}
\title{Compute precision and recall for all pairs in a pairwise_matrix}
\usage{
continuous_accuracy(distance_matrix, gold_standard)
}
\arguments{
\item{distance_matrix}{A symmetric matrix of distances}

\item{gold_standard}{A symmetric, binary matrix where 1 indicates a true pair}
}
\value{
A data.frame with columns 'precision' and 'recall'
}
\description{
Compute precision and recall for all pairs in a pairwise_matrix
}
