% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{conv_distance}
\alias{conv_distance}
\title{Find minimum distance between average MIDs from an MIData object}
\usage{
conv_distance(mi_data, x, y, e, distance, default = Inf)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{x}{peak index}

\item{y}{peak index}

\item{e}{experiment index}

\item{distance}{A similarity function on MIDs.}

\item{default}{Distance value to use when there is no possible convolution}
}
\value{
the resulting similarity value
}
\description{
Calculates the given similarity function between MIDs for peak x and y for
experiment e from an MIData object (mi_data).
}
