% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{conv_reduce_all}
\alias{conv_reduce_all}
\title{Same as conv_reduce, but for all pairs x*y, returns a matrix.
this avoids recalculating convolutions}
\usage{
conv_reduce_all(mi_data, e, f, g)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{e}{experiment index}

\item{f}{A function f(x, y) taking two MIDs.}

\item{g}{a function g taking a vector of values f1, f2, ...}
}
\value{
the matrix of g(f(x,y) ...) values for all x,y
}
\description{
Same as conv_reduce, but for all pairs x*y, returns a matrix.
this avoids recalculating convolutions
}
