% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{conv_reduce_block}
\alias{conv_reduce_block}
\title{compute conv_reduce for one matrix block where all x are the same size
and all y are the same size, but y is larger than x}
\usage{
conv_reduce_block(mi_data, e, f, g, x_index, y_index, z_index)
}
\arguments{
\item{mids_x}{matrix of MIDs for metabolites x, each column one MID}

\item{mids_y}{matrix of MIDs for metabolites y}

\item{mids_z}{matrix of MIDs for metabolites z, may be empty}
}
\description{
compute conv_reduce for one matrix block where all x are the same size
and all y are the same size, but y is larger than x
}
