% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{conv_similarity}
\alias{conv_similarity}
\title{Calculate similarity, using convolution}
\usage{
conv_similarity(sim_data, x, y, e, similarity, remove_m0 = F)
}
\arguments{
\item{sim_data}{the MIdata object}

\item{x}{peak index}

\item{y}{peak index}

\item{e}{experiment index}

\item{similarity}{similarity function}

\item{remove_m0}{whether to exclude M+0 (the first element in MID vectors) from similarity calculation}
}
\value{
the convoluted MID vector, or NA if isotopoic enrichment is less than than tolerance
}
\description{
brings MIDs of peak x and peak y for experiment e from an MIData object (sim_data)
computes convolution in case of unequal C numbers
calculates similarity between MIDs
removes the first element that corresponds to M+0, if remove_m0 == T, before computing similarity.
}
