% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{convolute}
\alias{convolute}
\title{Compute the convolution x*y of two MIDs x,y}
\usage{
convolute(x, y)
}
\arguments{
\item{x}{an MID vector}

\item{y}{an MID vector}
}
\value{
the convolution MID vector x*y
}
\description{
Compute the convolution x*y of two MIDs x,y
}
