% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{convolute_array}
\alias{convolute_array}
\title{Convolute columns of x with second dimension of y}
\usage{
convolute_array(x_mat, y_array)
}
\arguments{
\item{x_mat}{a matrix with MIDs in columns (MI x experiments)}

\item{y_array}{an MI x experiments x peaks array}
}
\value{
An MI x experiments x peaks array of convolutions
}
\description{
Convolute columns of x with second dimension of y
}
