% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{convolute_cols}
\alias{convolute_cols}
\title{Convolute x*y for all column vectors y in matrix y_mat}
\usage{
convolute_cols(x, y_mat)
}
\arguments{
\item{x}{an MID vector}

\item{y_mat}{a matrix whose columns are MID vectors}
}
\value{
the matrix of convolution vectors x*y for each column y in y_mat
}
\description{
Convolute x*y for all column vectors y in matrix y_mat
}
