% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{convolution_matrix}
\alias{convolution_matrix}
\title{Generate a convolution matrix for an MID}
\usage{
convolution_matrix(x, y_carbons)
}
\arguments{
\item{x}{an MID}

\item{y_carbons}{the number of carbons in the vector y}
}
\description{
Convolution of two MIDs x * y can be written as a matrix multiplication
A(x).y
This function creates the matrix A(x) for a given number of carbons in y
}
