% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{correct_and_remove}
\alias{correct_and_remove}
\title{Correct an MID vector for naturally occurring isotopes and remove M+0}
\usage{
correct_and_remove(mid)
}
\arguments{
\item{mid}{An MID vector to correct
#' @export}
}
\description{
Correct an MID vector for naturally occurring isotopes and remove M+0
}
