% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{correct_natural_13c}
\alias{correct_natural_13c}
\title{Perform 13C correction}
\usage{
correct_natural_13c(mi_data)
}
\arguments{
\item{mi_data}{An MIData object}
}
\value{
A new MIData object with 13C-corrected MIDs
}
\description{
Correct each MID for natural 13C abundance and return the corrected data.
}
