% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{cosine_dist}
\alias{cosine_dist}
\title{Cosine distance between two vectors of the same length.}
\usage{
cosine_dist(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the cosine similarity, or NA if either x or y is a zero vector
}
\description{
This is 1 minus the cosine similarity.
}
