% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{cosine_dist_no_m0}
\alias{cosine_dist_no_m0}
\title{Cosine distance between two vectors of the same length without M+0.}
\usage{
cosine_dist_no_m0(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the cosine similarity, or NA if either x or y is a zero vector
}
\description{
This is 1 minus the cosine similarity.
}
