% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Cosine similarity between two vectors of the same length.}
\usage{
cosine_sim(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the cosine similarity, or NA if either x or y is a zero vector
}
\description{
Cosine similarity between two vectors of the same length.
}
