% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{cosine_sim_no_m0}
\alias{cosine_sim_no_m0}
\title{Cosine similarity between two vectors of the same length without M+0.}
\usage{
cosine_sim_no_m0(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the cosine similarity, or NA if either x or y is a zero vector
}
\description{
Cosine similarity between two vectors of the same length without M+0.
}
