% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dot_dist}
\alias{dot_dist}
\title{Dot product distance (not really a distance)}
\usage{
dot_dist(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
a distance based on the dot (scalar) product x.y
This is always nonnegative, and zero if x = y
}
\description{
Dot product distance (not really a distance)
}
