% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dot_sim}
\alias{dot_sim}
\title{Dot product similarity}
\usage{
dot_sim(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the dot (scalar) product x.y
}
\description{
Dot product similarity
}
