% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{dot_sim_no_m0}
\alias{dot_sim_no_m0}
\title{Dot product similarity without M+0}
\usage{
dot_sim_no_m0(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the dot (scalar) product x.y
}
\description{
Dot product similarity without M+0
}
