% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{enrichment_dist_matrix}
\alias{enrichment_dist_matrix}
\title{Compute distance matrix from isotopic enrichment}
\usage{
enrichment_dist_matrix(midata, experiments, method = "euclidean")
}
\arguments{
\item{midata}{An MIData object}

\item{experiments}{A list of experiments (names) to use}

\item{method}{Name of  distance metric, as in stats::dist. Use either
of "euclidean", "manhattan", or "canberra"}
}
\description{
Compute distance matrix from isotopic enrichment
}
