% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{euclidean_dist}
\alias{euclidean_dist}
\title{The Euclidean distance}
\usage{
euclidean_dist(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the Euclidean distance between x and y
}
\description{
The Euclidean distance
}
