% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{euclidean_dist_no_m0}
\alias{euclidean_dist_no_m0}
\title{The Euclidean distance without M+0}
\usage{
euclidean_dist_no_m0(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the Euclidean distance between x and y without M+0
}
\description{
The Euclidean distance without M+0
}
