% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{euclidean_dist_sq}
\alias{euclidean_dist_sq}
\title{The squared Euclidean distance (sum of squares)}
\usage{
euclidean_dist_sq(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the square of the Euclidean distance between x and y
}
\description{
The squared Euclidean distance (sum of squares)
}
