% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{euclidean_dist_sq_no_m0}
\alias{euclidean_dist_sq_no_m0}
\title{The squared Euclidean distance (sum of squares) without M+0}
\usage{
euclidean_dist_sq_no_m0(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the square of the Euclidean distance between x and y
}
\description{
The squared Euclidean distance (sum of squares) without M+0
}
