% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{euclidean_sum_dist}
\alias{euclidean_sum_dist}
\title{Sum of euclidean distance on columns}
\usage{
euclidean_sum_dist(x, y)
}
\arguments{
\item{x}{a matrix}

\item{y}{a matrix}
}
\value{
sum of Euclidean distance between columns
}
\description{
A distance function for matrices. Takes the sum of the
Euclidean distances between each column of x, y.
}
