% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-arguments.R
\name{fetch_input_args}
\alias{fetch_input_args}
\title{Reads input_args from input_file_name and returns the input set for the row input_row_index}
\usage{
fetch_input_args(input_row_index, input_file_name)
}
\description{
Reads input_args from input_file_name and returns the input set for the row input_row_index
}
