% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{filter_enrichment}
\alias{filter_enrichment}
\title{Return a zero vector if the mid enrichment is below a threshold}
\usage{
filter_enrichment(mid, tol = 0.0107)
}
\arguments{
\item{mid}{An MID vector}

\item{tol}{A threshold value}
}
\value{
The given mid vector, or a vector of zeros of the same length,
if the isotopic enrichment is less than threshold
}
\description{
Return a zero vector if the mid enrichment is below a threshold
}
