% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{filter_pairwise_matrix}
\alias{filter_pairwise_matrix}
\title{Filters the pairwise distance matrix by only including the top percent distances from the CDF.
Filtering is performed for each row.
Assumes distances only.}
\usage{
filter_pairwise_matrix(pairwise_matrix, percentile = 0.01)
}
\arguments{
\item{pairwise_matrix}{a pairwise distance matrix}

\item{percentile}{top X percent of the predictions will be kept where X is 1 per default.}
}
\description{
Filters the pairwise distance matrix by only including the top percent distances from the CDF.
Filtering is performed for each row.
Assumes distances only.
}
