% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{find_mi_index}
\alias{find_mi_index}
\title{Compute index vector into the MI data table
for a given list of atom sizes}
\usage{
find_mi_index(peak_n_atoms)
}
\description{
Compute index vector into the MI data table
for a given list of atom sizes
}
