% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mid}
\alias{get_avg_mid}
\title{Get an averaged MID vector}
\usage{
get_avg_mid(mi_data, p, e)
}
\arguments{
\item{mi_data}{an MIData object}

\item{p}{the peak index}

\item{e}{the experiment index}
}
\description{
for a given peak.
}
