% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mid_all}
\alias{get_avg_mid_all}
\title{Get averaged MID vectors}
\usage{
get_avg_mid_all(mi_data, index)
}
\arguments{
\item{mi_data}{an MIData object}

\item{index}{the peak index}
}
\value{
a matrix where each column is the MID from an experiment
}
\description{
for a given peak index, for all experiments
}
