% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mids}
\alias{get_avg_mids}
\title{Get averaged MIDs for several peaks from an MIData object, each}
\usage{
get_avg_mids(mi_data, peak_index, e)
}
\arguments{
\item{mi_data}{an MIData object}

\item{peak_index}{One or more peak indices}

\item{e}{an experiment index}
}
\value{
An array of dimensions MI x experiments x peaks
}
\description{
Get averaged MIDs for several peaks from an MIData object, each
}
