% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_exp_indices}
\alias{get_exp_indices}
\title{Get indices into the columns of the $mid matrix for experiment e}
\usage{
get_exp_indices(mi_data, e)
}
\arguments{
\item{mi_data}{An MIData object}

\item{e}{An experiment index}
}
\description{
Get indices into the columns of the $mid matrix for experiment e
}
