% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{get_fun_index}
\alias{get_fun_index}
\title{Returns the metabolite index which was chosen by fun.
It is supposed to be used in the combine() function.}
\usage{
get_fun_index(vector, fun)
}
\arguments{
\item{vector}{a row from a pairwise matrix}

\item{fun}{a function like max() or min()}
}
\description{
Returns the metabolite index which was chosen by fun.
It is supposed to be used in the combine() function.
}
