% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold-standard.R
\name{get_global_percentile_accuracy}
\alias{get_global_percentile_accuracy}
\title{Compute accuracy measures at a given percentile cutoffs for pairwise_matrix}
\usage{
get_global_percentile_accuracy(
  pairwise_matrix,
  gold_standard,
  measure,
  noise,
  experiment,
  subset_size,
  percentiles
)
}
\arguments{
\item{pairwise_matrix}{A symmetric matrix of distances}

\item{gold_standard}{A symmetric, binary matrix where 1 indicates a true pair}

\item{measure}{?}

\item{noise}{?}

\item{experiment}{?}

\item{subset_size}{?}

\item{percentiles}{A vector of percentiles, in the range 0--100

#' @returns A data frame with accuracy values at each percentile}
}
\description{
Compute accuracy measures at a given percentile cutoffs for pairwise_matrix
}
