% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_lower_triangular}
\alias{get_lower_triangular}
\title{Get Vector of lower-triangular elements of a square matrix}
\usage{
get_lower_triangular(mat)
}
\arguments{
\item{mat}{a square matrix}
}
\value{
a vector of matrix elements below the diagonal
}
\description{
Get Vector of lower-triangular elements of a square matrix
}
