% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_mi_indices}
\alias{get_mi_indices}
\title{Get MI indices of a given peak}
\usage{
get_mi_indices(mi_data, p)
}
\arguments{
\item{mi_data}{an MIData object}

\item{p}{the peak index}
}
\value{
a vector of MI indices
}
\description{
Get MI indices of a given peak
}
