% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_mid_dist}
\alias{get_mid_dist}
\title{Apply a distance function dist to MID x,y after convolution}
\usage{
get_mid_dist(dist, x, y, conv_fn)
}
\arguments{
\item{dist}{a distance function}

\item{x}{an MID}

\item{y}{an MID}

\item{conv_fn}{a function used to convolute MIDs}
}
\value{
the distance value
}
\description{
if length(x) < length(y), we find a vector v such that
||x * v - y|| is minimal, and compute dist(x*v, y)
and vice versa if length(x) > length(y)
}
