% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_midata}
\alias{get_midata}
\title{Read peak areas from file and create an MIData object}
\usage{
get_midata(peak_areas_fname)
}
\arguments{
\item{peak_areas_fname}{Name of a a tab-separated file containing peak areas}
}
\value{
An MIData object
}
\description{
Read peak areas from file and create an MIData object
}
