% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_mids}
\alias{get_mids}
\title{Get MIDs for a given peak and experiment from an MIData object}
\usage{
get_mids(mi_data, p, e)
}
\arguments{
\item{mi_data}{an MIData object}

\item{p}{the peak index}

\item{e}{the experiment index}
}
\value{
A matrix with MIDs in columns
}
\description{
Get MIDs for a given peak and experiment from an MIData object
}
