% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_peak_index}
\alias{get_peak_index}
\title{Get the index of a list of peak identifers in an MIData object}
\usage{
get_peak_index(mi_data, peak_ids)
}
\arguments{
\item{mi_data}{an MIData object}

\item{peak_ids}{a list of peak identifiers}
}
\description{
Get the index of a list of peak identifers in an MIData object
}
