% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_peak_index_n_atoms}
\alias{get_peak_index_n_atoms}
\title{Get the index of peaks with a specific number of atoms}
\usage{
get_peak_index_n_atoms(mi_data, n_atoms)
}
\arguments{
\item{mi_data}{an MIData object}

\item{n_atoms}{number of atoms in peaks of interest}
}
\description{
Get the index of peaks with a specific number of atoms
}
