% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_peak_n_atoms}
\alias{get_peak_n_atoms}
\title{Get the number of atoms for a given peak}
\usage{
get_peak_n_atoms(mi_data, p)
}
\arguments{
\item{mi_data}{an MIData object}

\item{p}{the peak index}
}
\description{
Get the number of atoms for a given peak
}
