% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_distance_matrix}
\alias{is_distance_matrix}
\title{Test if a matrix is a valid distance matrix (metric)}
\usage{
is_distance_matrix(mat)
}
\arguments{
\item{mat}{A (candidate) distance matrix
NOTE: this function should be fixed, see https://github.com/Nilsson-Lab-KI/remn/issues/40}
}
\description{
Test if a matrix is a valid distance matrix (metric)
}
