% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{isotopic_enrichment}
\alias{isotopic_enrichment}
\title{Isotopic enrichment of an MID (fraction of heavy atoms)}
\usage{
isotopic_enrichment(mid)
}
\arguments{
\item{mid}{An MID vector}
}
\value{
The enrichment value
}
\description{
Isotopic enrichment of an MID (fraction of heavy atoms)
}
