% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{jensen_shannon}
\alias{jensen_shannon}
\title{The Jensen-Shannon (JS) distance}
\usage{
jensen_shannon(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the JS distance between x and y
}
\description{
The Jensen-Shannon (JS) distance
}
