% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{manhattan_distance}
\alias{manhattan_distance}
\title{The manhattan distance sum_i |x-i - y_i|}
\usage{
manhattan_distance(x, y)
}
\arguments{
\item{x}{a vector}

\item{y}{a vector}
}
\value{
the manhattan distance between x and y
}
\description{
The manhattan distance sum_i |x-i - y_i|
}
