% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{max_nonempty}
\alias{max_nonempty}
\title{Take maximum of x after discarding missing values,
return NA if the result is empty. If x has an "index"
attribute, it must be a list of same length as x;
the element corresponding to the maximum is added as
an "index" attribute to the result}
\usage{
max_nonempty(x)
}
\description{
NOTE: I'm not convinced we should return NA in
case of empty lists; see ?max for reasons for using +/-Inf
Keeping this for compatibility for now
}
