% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formulas.R
\name{merge_formula}
\alias{merge_formula}
\title{Takes as input a broken formula and merges it back to a character string}
\usage{
merge_formula(broken_formula)
}
\description{
Takes as input a broken formula and merges it back to a character string
}
