% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{mi_weighted_dist_normalized}
\alias{mi_weighted_dist_normalized}
\title{A p-distance weighted by the mass isotopomer number, normalized
Here, the weight is i/n where i is the MI number and n is the number of atoms.}
\usage{
mi_weighted_dist_normalized(x, y, p = 1)
}
\arguments{
\item{x}{an MID vector}

\item{y}{an MID vector}

\item{p}{The exponent p in the p-norm}
}
\value{
the distance between x and y
}
\description{
A p-distance weighted by the mass isotopomer number, normalized
Here, the weight is i/n where i is the MI number and n is the number of atoms.
}
