% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance-measures.R
\name{mi_weighted_distance}
\alias{mi_weighted_distance}
\title{A p-distance weighted by the mass isotopomer number}
\usage{
mi_weighted_distance(x, y, p = 1)
}
\arguments{
\item{x}{an MID vector}

\item{y}{an MID vector}

\item{p}{The exponent p in the p-norm}
}
\value{
the distance between x and y
}
\description{
A p-distance weighted by the mass isotopomer number
}
