% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{midata_subset}
\alias{midata_subset}
\title{Subset an MIData object to the peaks given by peak_index, and return a new MIData object}
\usage{
midata_subset(mi_data, peak_subset_index)
}
\arguments{
\item{mi_data}{MIData object}

\item{peak_subset_index}{indices of peaks to be included in the subset}
}
\description{
Subset an MIData object to the peaks given by peak_index, and return a new MIData object
}
