% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{midata_transform}
\alias{midata_transform}
\title{Apply a function to each MID in an MIData object}
\usage{
midata_transform(midata, f)
}
\arguments{
\item{midata}{An MIData object}

\item{f}{The function to apply. Must take an MID vector as first argument
and return a valid MID vector of the same size.}
}
\description{
This can be used for example to apply 13C correction to a data set
}
