% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remn-package.R
\docType{package}
\name{midist-package}
\alias{midist}
\alias{midist-package}
\title{midist: MID distance measures}
\description{
Tools for computing distance (or similarity) measures between mass isotopomer distributions (MIDs).
}
\author{
\strong{Maintainer}: Deniz Secilmis \email{deniz.secilmis@ki.se} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Roland Nilsson \email{roland.nilsson@ki.se} (\href{https://orcid.org/0000-0002-6020-7498}{ORCID})
}

}
\keyword{internal}
