% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{misplace_peak_ids}
\alias{misplace_peak_ids}
\title{Reorder the peak IDs of an mi_data object}
\usage{
misplace_peak_ids(midata)
}
\arguments{
\item{midata}{An MIData object}
}
\value{
A vector of reordered peak IDs
}
\description{
Reorder the peak IDs of an mi_data object
}
