% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\docType{data}
\name{natural_13C_fraction}
\alias{natural_13C_fraction}
\title{Constant for the fraction of naturally occurring 13C atoms}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
natural_13C_fraction
}
\description{
Constant for the fraction of naturally occurring 13C atoms
}
\keyword{datasets}
