% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{natural_mid}
\alias{natural_mid}
\title{Generate a natural (binomial) MID}
\usage{
natural_mid(n, p = natural_13C_fraction)
}
\arguments{
\item{n}{number of atoms}

\item{p}{heavy isotope probability}
}
\value{
an MID vector
}
\description{
Generate a natural (binomial) MID
}
