% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pairwise_matrix_all_ds}
\alias{pairwise_matrix_all_ds}
\title{Calculate similarity matrix, based on parameters specified in an InputData object}
\usage{
pairwise_matrix_all_ds(e, input_data)
}
\arguments{
\item{e}{experiment index that corresponds to the same index in MIData}

\item{input_data}{InputData object (for more detail see ...)}
}
\description{
Calculates a similarity matrix based on the given similarity function for all peak pairs
for experiment e from an MIData object (midata).
If a list of reactions is input to input_data$reaction_data, the given similarity function will only be applied to peak pairs
whose formula or mass difference (depending on "input_data$reaction_restriction") can be matched an allowed reaction
}
