% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{pairwise_matrix_ds}
\alias{pairwise_matrix_ds}
\title{Calculate similarity matrix, based on parameters specified in an InputData object}
\usage{
pairwise_matrix_ds(e, input_data, write_to_file = F, return = T)
}
\arguments{
\item{e}{experiment index that corresponds to the same index in MIData}

\item{input_data}{InputData object (for more detail see ...)}

\item{write_to_file}{TRUE writes the similarity matrix to file. FALSE by default}

\item{return}{whether to return the similarity matrix. Choose FALSE for commandline runs}
}
\description{
Calculates a similarity matrix based on the given similarity function for all peak pairs
for experiment e from an MIData object (midata).
If a list of reactions is input to input_data$reaction_data, the given similarity function will only be applied to peak pairs
whose formula or mass difference (depending on "input_data$reaction_restriction") can be matched an allowed reaction
}
