\name{plot_density}
\alias{plot_density}
\title{
Plot Density
}
\description{
Plots probability density functions
}
\usage{
plot_density(...)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
}

\value{
Outputs a probability density function plot for each PDF given
}
