\name{plot_density.wo}
\alias{plot_density.wo}
\title{
Plot Density write out
}
\description{
Plots probability density functions to file
}
\usage{
plot_density.wo(..., file)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
  \item{file}{
output file
}
}

\value{
Outputs a probability density function plot to PDF file
}
