\name{plot_density_log}
\alias{plot_density_log}

\title{
Plot Density log
}
\description{
Plots probability density functions in semilog plot
}
\usage{
plot_density_log(...)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
}

\value{
Outputs a probability density function plot for each PDF given. Semilog plot
}
