\name{plot_density_log.wo}
\alias{plot_density_log.wo}

\title{
Plot Density log
}
\description{
Plots probability density functions in semilog plot
}
\usage{
plot_density_log.wo(..., file)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
  \item{file}{
output file
}
}

\value{
PDF file of a probability density function plot for each PDF given. Semilog plot
}
