\name{plot_density_zoom}
\alias{plot_density_zoom}

\title{
Plot Density Zoom
}
\description{
Plots probability density functions, zoomed to certain interval
}
\usage{
plot_density_zoom(..., zoom)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
  \item{zoom}{
fraction of x axis to zoom
}

}

\value{
Outputs a probability density function plot for each PDF given
}
