\name{plot_density_zoom.wo}
\alias{plot_density_zoom.wo}

\title{
Plot Density Zoom
}
\description{
Plots probability density functions, zoomed to certain interval
}
\usage{
plot_density_zoom.wo(..., zoom, file)
}

\arguments{
  \item{\dots}{
Objects of class density as produced by density()
}
  \item{zoom}{
fraction of x axis to zoom
}
  \item{file}{
output file
}
}
\value{
PDF file of a probability density function plot for each PDF given. Semilog plot
}
