% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mid_barchart}
\alias{plot_mid_barchart}
\title{Plot MID as barchart with error bars}
\usage{
plot_mid_barchart(mid_mat)
}
\arguments{
\item{mid_mat}{A matrix whose columns are MID vectors}
}
\value{
A ggplot object
}
\description{
Plot MID as barchart with error bars
}
