% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{random_dirichlet}
\alias{random_dirichlet}
\title{Sample from a Dirichlet distribution.}
\usage{
random_dirichlet(mean, precision, n)
}
\arguments{
\item{mean}{Mean vector of the Dirichlet distribution. Must sum to 1.}

\item{precision}{Precision of the Dirichlet distribution,}

\item{n}{number of samples to draw}
}
\value{
A matrix with n columns, each an MID of the same length as mean
}
\description{
Draws n samples from a Dirichlet distribution with the given mean vector
and precision. Each component i of the resulting random vector has
marginal variance x_i*(1-x_i) / (1 + precision) where x_i is the i'th component
of the mean vector.
}
