% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{random_mid}
\alias{random_mid}
\title{Sample MIDs from a Dirichlet distributon.}
\usage{
random_mid(mean, stdev, n)
}
\arguments{
\item{mean}{An MID vector}

\item{stdev}{A standard deviation}

\item{n}{number of samples to draw}
}
\value{
A matrix with n columns, each an MID of the same length as mean
}
\description{
Samples n MIDs from a Dirichlet distribution with the given mean vector
and a precision such that the standard deviation of an MI fraction
with (marginal) expectation = 1/2 equals stddev.
}
