% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remn-package.R
\docType{package}
\name{remn-package}
\alias{remn}
\alias{remn-package}
\title{remn: Reverse Engineering metabolic networks}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
