% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{remn_v1}
\alias{remn_v1}
\title{Compute pairwise distance matrices for individual experiments}
\usage{
remn_v1(midata, f, g_select, rdata_fname, return = T)
}
\arguments{
\item{midata}{An MIData object}

\item{f}{A distance function, as in conv_reduce}

\item{g_select}{A function for selecting best convolutions, as in conv_reduce}

\item{rdata_fname}{NOT USED}

\item{return}{NOT USED}
}
\value{
A list of distance matrices
}
\description{
Compute pairwise distance matrices for individual experiments
}
