% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{remn_v2}
\alias{remn_v2}
\title{Compute a pairwise distance matrix and optionally save it}
\usage{
remn_v2(midata, f, g_select, rdata_fname, return = T)
}
\arguments{
\item{midata}{An MIData object}

\item{f}{A distance function}

\item{g_select}{A selection function}

\item{rdata_fname}{A file name to save results in, if return == F}

\item{return}{A boolean; if FALSE, results are saved using base::save}
}
\value{
A pairwise matrix if return == TRUE; else nothing
}
\description{
Compute a pairwise distance matrix and optionally save it
}
