% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{replace_na_with_max}
\alias{replace_na_with_max}
\title{Replace NAs in an array with the maximal value}
\usage{
replace_na_with_max(x)
}
\arguments{
\item{x}{an array}
}
\value{
the array with NA values replaced
}
\description{
Replace NAs in an array with the maximal value
}
