% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{similarity_matrix}
\alias{similarity_matrix}
\title{Calculate similarity matrix, based on a specific similarity/distance measure, for metabolites from an MIData object}
\usage{
similarity_matrix(midata, e, similarity_measure, remove_m0 = FALSE)
}
\arguments{
\item{midata}{the MIdata object}

\item{e}{experiment index}

\item{similarity_measure}{A similarity function on MIDs}

\item{remove_m0}{whether to remove M+0 from MIDs (TRUE if MIDs are corrected for the natural 13C. FALSE by default)}
}
\value{
the resulting similarity matrix
}
\description{
Calculates a similarity matrix based on the given similarity function for all MID pairs for
experiment e from an MIData object (midata).
}
