% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{similarity_matrix_with_reaction_filter}
\alias{similarity_matrix_with_reaction_filter}
\title{Calculate similarity matrix, IF THE FORMULA OR MASS DIFFERENCE IS ACCEPTABLE,
based on a specific similarity/distance measure, for metabolites from an MIData object}
\usage{
similarity_matrix_with_reaction_filter(
  midata,
  e,
  similarity_measure,
  allowed_reactions,
  type = "formula",
  remove_m0 = FALSE,
  tol = 10
)
}
\arguments{
\item{midata}{the MIdata object}

\item{e}{experiment index}

\item{similarity_measure}{A similarity function on MIDs}

\item{allowed_reactions}{a vector of allowed reactions. Either formulas or masses depending on whether type is "formulas" or "masses", respectively}

\item{type}{how to asses the intermediate reaction - based on whether formula or mass.}

\item{remove_m0}{whether to remove M+0 from MIDs (TRUE if MIDs are corrected for the natural 13C. FALSE by default)}
}
\value{
the resulting similarity matrix
}
\description{
Calculates a similarity matrix based on the given similarity function for all peak pairs
whose formula or mass difference can be matched an allowed reaction
for experiment e from an MIData object (midata).
}
