% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{[<-.AssignList}
\alias{[<-.AssignList}
\title{Operator for assignment to an AssignList}
\usage{
\method{[}{AssignList}(assign_list, ...) <- value
}
\arguments{
\item{assign_list}{the unique object assign_list from this package}

\item{...}{elements to assign}

\item{value}{the list to assign from}
}
\value{
the assign_list object
}
\description{
For example, assign_list[x, y] <- list(a,b)  is effectively the same as
x <- a and y <- b.
See help for "[<-"]
}
