% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{weight_pm_by_enrichment}
\alias{weight_pm_by_enrichment}
\title{Weighs the pairwise measure by the enrichment scores of metabolites involved.
Assumes distances only.
In case of equal C, the pairwise distance is divided by isotopic_enrichment(mid_x)*isotopic_enrichment(mid_y) for pair (x,y)
In case of unequal C, it is divided by isotopic_enrichment(mid_x)*isotopic_enrichment(mid_z)\emph{isotopic_enrichment(mid_y) for pair (x}z,y)
Returns a penalized pairwise distance matrix.}
\usage{
weight_pm_by_enrichment(
  pairwise_matrix,
  middle_met_matrix,
  experiment_matrix,
  mi_data
)
}
\arguments{
\item{pairwise_matrix}{distance matrix}

\item{middle_met_matrix}{matrix of middle metabolites that were chosen as the best convolutions}

\item{experiment_matrix}{matrix of the same size, whose cells denote the experiment that was the output of g_select()}

\item{mi_data}{the MIData object}
}
\description{
Weighs the pairwise measure by the enrichment scores of metabolites involved.
Assumes distances only.
In case of equal C, the pairwise distance is divided by isotopic_enrichment(mid_x)*isotopic_enrichment(mid_y) for pair (x,y)
In case of unequal C, it is divided by isotopic_enrichment(mid_x)*isotopic_enrichment(mid_z)\emph{isotopic_enrichment(mid_y) for pair (x}z,y)
Returns a penalized pairwise distance matrix.
}
