% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{with_attr}
\alias{with_attr}
\title{Get object with an attribute added}
\usage{
with_attr(x, attr_name, attr_value)
}
\arguments{
\item{x}{Any object}

\item{attr_name}{name of the attribut to add}

\item{attr_value}{the attribute value to add}
}
\value{
The object x with attr(x, attr_name) == attr_value
}
\description{
Get object with an attribute added
}
