% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{without_attr}
\alias{without_attr}
\title{Get object without attributes}
\usage{
without_attr(x)
}
\arguments{
\item{x}{Any object}
}
\value{
The object x with attributes(x) == NULL
}
\description{
Get object without attributes
}
