% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{add_gamma_noise}
\alias{add_gamma_noise}
\title{Add multiplicative gamma noise to a given MID from a given standard deviation}
\usage{
add_gamma_noise(mid, stdev)
}
\arguments{
\item{mid}{An MID vector}

\item{stdev}{standard deviation for the gamma distribution
Computes the gamma distribution parameters theta and k from the mean of the MID and input standard deviation.
Multiplies the gamma values by the MID itself (multiplicative)
Adds this noise to the MID and returns the noisy MID}
}
\description{
Add multiplicative gamma noise to a given MID from a given standard deviation
}
