% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{c13correct}
\alias{c13correct}
\title{Correct an MID vector for naturally occurring isotopes}
\usage{
c13correct(mid, constraint = TRUE)
}
\arguments{
\item{mid}{An MID vector to correct}

\item{constraint}{NOT CURRENTLY USED}
}
\description{
Correct an MID vector for naturally occurring isotopes
}
