% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{c13correct}
\alias{c13correct}
\title{Correct an MID vector for naturally occurring isotopes}
\usage{
c13correct(mid, p = 0.0107, constraint = TRUE)
}
\arguments{
\item{mid}{An MID vector to correct}

\item{p}{The heavy atom natural abundance}

\item{constraint}{whether to constrain sum of corrected vector to 1}
}
\description{
Correct an MID vector for naturally occurring isotopes
}
