% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{conv_reduce}
\alias{conv_reduce}
\title{Calculates f(x, y) between MIDs for peak x and y for
experiment e from an MIData object (mi_data)
If x,y have unequal atom numbers, computes f(x*z, y) for all atoms of the same
size as y (if y is the longer vector) and returns g( ... ) of the list of
value of f. For example f = max gives the maximum value. The function g
must handle empty lists g(c()) in case there are no possible convolutions.}
\usage{
conv_reduce(mi_data, x, y, e, f, g)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{x}{peak index}

\item{y}{peak index}

\item{e}{experiment index}

\item{f}{A function f(x, y) taking two MIDs.}

\item{g}{a function g taking a vector of values f1, f2, ...}
}
\value{
the resulting value g(f1, f2, ...)
}
\description{
Calculates f(x, y) between MIDs for peak x and y for
experiment e from an MIData object (mi_data)
If x,y have unequal atom numbers, computes f(x*z, y) for all atoms of the same
size as y (if y is the longer vector) and returns g( ... ) of the list of
value of f. For example f = max gives the maximum value. The function g
must handle empty lists g(c()) in case there are no possible convolutions.
}
