% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{conv_reduce}
\alias{conv_reduce}
\title{Convolution-based measure between MIDs for peak x and y for
experiment e from an MIData object (mi_data).}
\usage{
conv_reduce(mi_data, x, y, e, f, g, impute = FALSE)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{x}{peak index}

\item{y}{peak index}

\item{e}{experiment index}

\item{f}{A function f(x, y) taking two MIDs and returning a scalar.}

\item{g}{a function g taking a non-empty vector of values and returning
the index of the "best" element; for example, which.min}

\item{impute}{whether to impute missing convolutants (prevent NA distances)}
}
\value{
the resulting value g(f1, f2, ...)
}
\description{
If x, y have the same number of atoms, f(x,y) is returned.
If x is smaller than y, this function computes f(x\emph{z, y) for all
peaks z such that x}z is the same size as y (and vice versa), and calls g()
on the resulting vector of values from f to select a value.
For example, setting g = which.max gives the maximum value.
}
