% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{conv_reduce_all}
\alias{conv_reduce_all}
\title{Same as conv_reduce, but calculates f(x, y) pairwise for all peaks x and y
from an MIData object (mi_data)}
\usage{
conv_reduce_all(mi_data, e, f, g)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{e}{Experiment index. If e is a vector, the function f(x, y)
must accept matrices.}

\item{f}{A function f(x, y) taking two MIDs, or two matrices whose columns are MIDs.}

\item{g}{a function g taking a vector of values f1, f2, ...}
}
\value{
the matrix of g(f(x,y) ...) values for all x,y
}
\description{
The function g must handle empty lists g(c()) in case there are no possible convolutions.
In that case, g_select will not be applied.
}
