% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convolution-distance.R
\name{conv_reduce_block}
\alias{conv_reduce_block}
\title{compute conv_reduce for one matrix block where all x are the same size
and all y are the same size, but y is larger than x}
\usage{
conv_reduce_block(mi_data, e, f, g, x_index, y_index, z_index)
}
\arguments{
\item{mi_data}{An MIData object}

\item{e}{experiment index}

\item{f}{A distance function}

\item{g}{A selection function, such as min_nonempty}

\item{x_index}{index of the peak x}

\item{y_index}{index of the peak y}

\item{z_index}{index of the peak z}
}
\description{
compute conv_reduce for one matrix block where all x are the same size
and all y are the same size, but y is larger than x
}
