% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{conv_similarity}
\alias{conv_similarity}
\title{Calculate similarity between average MIDs from an MIData object}
\usage{
conv_similarity(mi_data, x, y, e, similarity)
}
\arguments{
\item{mi_data}{the MIdata object}

\item{x}{peak index}

\item{y}{peak index}

\item{e}{experiment index}

\item{similarity}{A similarity function on MIDs.}
}
\value{
the resulting similarity value
}
\description{
Calculates the given similarity function between MIDs for peak x and y for
experiment e from an MIData object (mi_data).
If x,y have unequal atom numbers, computes all convolutions x*z of the same
size as y (if y is the longer vector) and returns the largest similarity;
if similarity returns NA for some pair, the NA value is ignored when taking
maximum similarity; if there are no possible convolutions, returns 0
}
