% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{create_atom_index}
\alias{create_atom_index}
\title{Create an index list mapping each number of atoms n to the indices of the peaks
having n atoms}
\usage{
create_atom_index(peak_n_atoms)
}
\arguments{
\item{peak_n_atoms}{number of atoms per peak}
}
\description{
NOTE: this is a more generic function, applicable to any list,
could be moved to util.R
}
