% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mid.R
\name{find_convolution}
\alias{find_convolution}
\title{Find the optimal convolution for two MIDs}
\usage{
find_convolution(z, x, tol = natural_13C_fraction)
}
\arguments{
\item{z}{the longer MID}

\item{x}{the shorter MID}

\item{tol}{a threshold below which the convolution is not computed}
}
\value{
the convoluted MID vector, or NA if isotopic enrichment is
less than than tolerance
}
\description{
Computes the convolution x\emph{y (for x}y = z), where x and z are the shorter and
longer MIDs, respectively, and y is the unknown MID
}
