% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mid}
\alias{get_avg_mid}
\title{Get averaged MIDs from an MIData object}
\usage{
get_avg_mid(mi_data, peak, exp)
}
\arguments{
\item{mi_data}{an MIData object}

\item{peak}{A peak index (integer) or identifier (string)}

\item{exp}{An optional experiment index (integer) or identifier (string)}
}
\value{
An MID vector, or, if exp is not a scalar, a matrix whose columns are MIDs
}
\description{
Retrieve a vector of averaged MIDs for a given peak and experiment,
or a matrix of MIDs for a given peak across several experiments.
}
