% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mid}
\alias{get_avg_mid}
\title{Get averaged MIDs from an MIData object}
\usage{
get_avg_mid(mi_data, p, e)
}
\arguments{
\item{mi_data}{an MIData object}

\item{p}{the peak index}

\item{e}{the experiment index (optional)}
}
\value{
An MID vector, or, if e is not a scalar, a matrix whose columns are MIDs
}
\description{
Retrieve a vector of averages MIDs for a peak p and an experiment e;
or, if e is a vector or is omitted, the matrix of MIDs for peak p across
the indicated experiments.
}
