% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{get_avg_mids}
\alias{get_avg_mids}
\title{Get averaged MIDs for several peaks from an MIData object}
\usage{
get_avg_mids(mi_data, peaks, exp)
}
\arguments{
\item{mi_data}{an MIData object}

\item{peaks}{One or more peak IDs or indices}

\item{exp}{One or more experiments IDs or indices}
}
\value{
For one peak, a matrix of dimensions MI x peaks;
for several peaks with same atom number, an array of dimensions
MI x experiments x peaks
}
\description{
Retrieve an array of averaged MIDs for a given peak and experiment,
or a matrix of MIDs for a given peak across several experiments.
}
