% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gold-standard.R
\name{get_continuous_accuracy}
\alias{get_continuous_accuracy}
\title{Compute precision and recall for all pairs in a pairwise_matrix}
\usage{
get_continuous_accuracy(
  pairwise_matrix,
  gold_standard,
  measure,
  subset_size,
  subset_sample_no,
  noise,
  noise_sample_no,
  experiment
)
}
\arguments{
\item{pairwise_matrix}{A symmetric matrix of distances}

\item{gold_standard}{A symmetric, binary matrix where 1 indicates a true pair}

\item{measure}{?}

\item{subset_size}{?}

\item{subset_sample_no}{?}

\item{noise}{?}

\item{noise_sample_no}{?}

\item{experiment}{?}
}
\description{
Compute precision and recall for all pairs in a pairwise_matrix
}
