% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is_distance_matrix}
\alias{is_distance_matrix}
\title{Test if a matrix is a valid distance matrix (metric)}
\usage{
is_distance_matrix(mat)
}
\arguments{
\item{mat}{A candidate distance matrix}
}
\value{
A list of string describing failure reasons
}
\description{
This checks that the matrix is symmetric, positive, has zero diagonal,
and obeys the triangle inequality;
for all triples (i, j, k) we must have A_ij + A_jk <= A_ik
}
