% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{max_nonempty}
\alias{max_nonempty}
\title{Take maximum of x after discarding missing values}
\usage{
max_nonempty(x)
}
\arguments{
\item{x}{A vector of values. If x has an "index"
attribute, it must be a vector of same length as x.}
}
\value{
The largest value in x, or NA if x is empty. If x has an "index"
attribute, the element corresponding to the largest value is added as
an "index" attribute to the result.
}
\description{
Take maximum of x after discarding missing values
}
