% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midata.R
\name{MIData}
\alias{MIData}
\title{Construct an MIData (mass isotopomer data) object}
\usage{
MIData(peak_areas, exp_names = NULL)
}
\arguments{
\item{peak_areas}{a peak_area data.frame. The first column of peak_areas
must be peak identifiers, repeated for each MI of the same peak.
The second column holds formulas (but )}

\item{exp_names}{a optional list of experiment names matching columns 2,3... in peak_areas}
}
\description{
Construct an MIData (mass isotopomer data) object
}
