\name{midist-package}
\alias{midist-package}
\alias{midist}
\docType{package}
\title{
\packageTitle{midist}
}
\description{
\packageDescription{midist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{midist}
\packageIndices{midist}
Functions to compare mutual information distributions in gene regulatory networks or similar
}
\author{
\packageAuthor{midist}

Maintainer: \packageMaintainer{midist}
}
\keyword{ package}

\examples{
#U_test(pdf.sif(sif_object), pdf.sif(sif_other)
}
