\name{midist-package}
\alias{midist-package}
\alias{midist}
\docType{package}
\title{
\packageTitle{midist}
}
\description{
\packageDescription{midist}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{midist}
\packageIndices{midist}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{midist}

Maintainer: \packageMaintainer{midist}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
