% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_convolution}
\alias{plot_convolution}
\title{Plot MID matrices for a convolution x * z = y}
\usage{
plot_convolution(mi_data, index_x, index_y, index_z)
}
\arguments{
\item{mi_data}{An MIData object}

\item{index_x}{index of peak x}

\item{index_y}{index of peak y}

\item{index_z}{index of peak z}
}
\value{
A ggplot object
}
\description{
Plot MID matrices for a convolution x * z = y
}
