% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_matrix}
\alias{plot_matrix}
\title{Plot a matrix as a heatmap}
\usage{
plot_matrix(mat)
}
\arguments{
\item{mat}{Any real matrix}
}
\value{
A ggplot object
}
\description{
Plot a matrix as a heatmap
}
